Minor changes in RecursiveMultiply

+++ Problem Sets/RecursiveMultiply.java
@@ -13,10 +13,10 @@ public class RecursiveMultiply{
 	public static void main(String[] args){	
 		
 		Scanner scan = new Scanner(System.in);
-		int a = Integer.parseInt(scan.next());
-		int b = Integer.parseInt(scan.next());
+		long a = Integer.parseInt(scan.next());
+		long b = Integer.parseInt(scan.next());
 		
-		long product = mult((long)a, (long)b);
+		long product = mult(a, b);
 		System.out.println(product);
 	}
